% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_checkReadMe.R
\name{da_checkReadMe}
\alias{da_checkReadMe}
\title{Check a ReadMe in a data archive for data applicants.}
\usage{
da_checkReadMe(AID, path = NULL, name = NULL, fstruct = 1)
}
\arguments{
\item{AID}{Character string of the application ID in the format of 'YYMM-DDx'}

\item{path}{Character string of the path to the ReadMe file. If \code{NULL},
the standard folder structure (see details/ \code{struct}) will be assumed.}

\item{fstruct}{Numeric value corresponding to the standard folder structure described in the
details section.}
}
\value{
Returns a list with the following entries:
\tabular{ll}{
files_not_in_dir \tab Which files are in the ReadMe but not in the directory (incl. subdirectories)? \cr
files_not_in_rm \tab Which files are in the (sub-) directories but not in the ReadMe?}
}
\description{
Check if the ReadMe and the list of files (and their names) of a data archive
correspond.
}
\details{
The ReadMe should list all files that are part of the data archive with their exact
names. At the same time, the data archive should include all files listed in the ReadMe.
\subsection{Standard ReadMe name and folder structure}{

The following standard folder structures (with the ReadMe situated in the last folder) are supported:
\tabular{rl}{1 \tab Q:/FDZ/Alle/02_Antragsverfahren/Verfahren_personalisiert/_2025/2501-01a_Mustermensch/Daten_2025-01-02/2501-01a_SUF_Off-site/Readme_2501-01a_PISA2000.txt}
To add new standard folder structures, please contact the package/function author. Alternatively,
please set the specific \code{path} in this function.
}
}
\examples{
# tbd

}
